requirement.txt
streamlit
langchain
langchain-community
langchain-openai
chromadb
redis
python-dotenv
